Periodically close open `request.Sessions` to avoid buggy interaction with Docker Desktop (#478)
* Periodically refresh open `requests.Session`s to mitigate open filehandle issues (#179)

As reported, we create a `requests.Session` object on first request to the servers and then reuse it indefinitely. This can leave some open file handles on the OS (not a big deal), but can interact poorly with a bug in Docker Desktop which causes the SDK to entierly break connections to the server.
See https://github.com/openai/openai-python/issues/140 for more info.

The order of items in the API responses is intentional, and this order is clobbered by the rendering of `OpenAIObject`. This change removes the alphabetic sort of response keys